chore: add initial Docker and environment configuration files
- Created .dockerignore to exclude unnecessary files from Docker builds.
- Added .env and .env.production files for environment variable management.
- Introduced docker-compose files for production and development setups.
- Created Dockerfile for application containerization.
- Updated various modules and services to enhance functionality and maintainability.

+++ src/modules/system/log/entities/task-log.entity.ts
@@ -7,7 +7,7 @@ import { TaskEntity } from '../../task/task.entity'
 
 @Entity({ name: 'sys_task_log' })
 export class TaskLogEntity extends CommonEntity {
-  @Column({ type: 'tinyint', default: 0 })
+  @Column({ type: 'smallint', default: 0 })
   @ApiProperty({ description: 'Task status: 0 failed, 1 success' })
   status: number
